% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3cache.R
\name{fetch_from_cache}
\alias{fetch_from_cache}
\title{Helper function for fetching a file from a cache directory.}
\usage{
fetch_from_cache(key, cache_dir)
}
\arguments{
\item{key}{character. The key under which the cache entry is stored.}

\item{cache_dir}{character. The cache directory. The default is
\code{cache_directory()}.}
}
\value{
the cached object if the cache has not invalidated. Otherwise,
  return \code{s3mpi::not_cached}.
}
\description{
This function will also test to determine whether the file has been
modified on S3 since the last cache save. If the file has never been
cached or the cache is invalidated, it will return \code{s3mpi::not_cached}.
}

