% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grab_latest_file_in_s3_dir.r
\name{grab_latest_file_in_s3_dir}
\alias{grab_latest_file_in_s3_dir}
\title{Find the latest modified file with a given S3 prefix.}
\usage{
grab_latest_file_in_s3_dir(path = s3path())
}
\arguments{
\item{path}{character. The S3 prefix to search for the
latest uploaded key.}
}
\description{
Find the latest modified file with a given S3 prefix.
}
\note{
This helper function is used when no input key
   is specified to the \code{\link{s3read}} function to fetch the
   latest modified file in that bucket. If the bucket has
   many files, this can be very slow.
}

