% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3cache.R
\name{last_modified}
\alias{last_modified}
\title{Determine the last modified time of an S3 object.}
\usage{
last_modified(key)
}
\arguments{
\item{key}{character. The s3 key of the object.}
}
\value{
the last modified time or \code{NULL} if it does not exist on S3.
}
\description{
Determine the last modified time of an S3 object.
}

