% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.get.R
\name{print.s3mpi_error}
\alias{print.s3mpi_error}
\title{Printing for s3mpi errors.}
\usage{
\method{print}{s3mpi_error}(x, ...)
}
\arguments{
\item{x}{ANY. R object to print.}

\item{...}{additional objects to pass to print function.}
}
\description{
Printing for s3mpi errors.
}
