% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3.get.R
\name{s3.get}
\alias{s3.get}
\title{Fetch an R object from an S3 path.}
\usage{
s3.get(bucket, bucket.location = "US", verbose = FALSE, debug = FALSE)
}
\arguments{
\item{bucket}{character. A full S3 path.}

\item{bucket.location}{character. Usually \code{"US"}.}

\item{verbose}{logical. If \code{TRUE}, the \code{s3cmd}
utility verbose flag will be set.}

\item{debug}{logical. If \code{TRUE}, the \code{s3cmd}
utility debug flag will be set.}
}
\value{
The R object stored in RDS format on S3 in the \code{bucket} path.
}
\description{
Fetch an R object from an S3 path.
}

