% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3cache.R
\name{s3cache}
\alias{s3cache}
\title{A caching layer around s3mpi calls.}
\usage{
s3cache(s3key, value)
}
\arguments{
\item{s3key}{character. The full S3 key to attempt to read or write
to the cache.}

\item{value}{ANY. The R object to save in the cache. If missing,
a cache read will be performed instead.}
}
\description{
Fetching large files from the S3 MPI can be expensive when performed
multiple times. This method allows one to add a caching layer
around S3 fetching. The user should specify the configuration option
\code{options(s3mpi.cache = "some/dir")}. The recommended cache
directory (where files will be stored) is \code{"~/.s3cache"}.
}

