% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3delete.R
\name{s3delete}
\alias{s3delete}
\title{Delete an R object from S3 by key}
\usage{
s3delete(key, path = s3path())
}
\arguments{
\item{key}{character. The key to delete from S3.}

\item{path}{character. The location of your S3 bucket as a prefix to \code{name},
for example, \code{"s3://mybucket/"} or \code{"s3://mybucket/myprefix/"}.}
}
\description{
Delete an R object from S3 by key
}
\seealso{
\code{\link{s3store}}
}

