% Generated by roxygen2 (4.0.2): do not edit by hand
\name{s3exists}
\alias{s3exists}
\title{Determine whether object exists on S3}
\usage{
s3exists(name, .path = s3path(), ...)
}
\arguments{
\item{name}{string. Name of file to look for}

\item{path}{string. Path to file}
}
\description{
Test whether or not the given object exists at the
give S3 path
}

