% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/s3read.r
\name{s3read}
\alias{s3read}
\title{Read an R object in S3 by key}
\usage{
s3read(name = NULL, .path = s3path(), ...)
}
\description{
Any type of object that can be serialized as an RDS file
is capable of being stored using this interface.
}
\examples{
\dontrun{
s3store(c(1,2,3), 'test123')
print(s3read('test123'))
# [1] 1 2 3
}
}

