% Generated by roxygen2 (4.0.2): do not edit by hand
\name{s3store}
\alias{s3store}
\title{Store an R object in S3 by key}
\usage{
s3store(obj, name = NULL, .path = s3path(), ...)
}
\description{
Any type of object that can be serialized as an RDS file
is capable of being retrieved using this interface.
}
\examples{
\dontrun{
s3store(c(1,2,3), 'test123')
print(s3read('test123'))
# [1] 1 2 3
}#'
}

