% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3cache.R
\name{save_to_cache}
\alias{save_to_cache}
\title{Helper function for saving a file to a cache directory.}
\usage{
save_to_cache(key, value, cache_dir = cache_directory())
}
\arguments{
\item{key}{character. The key under which the cache entry is stored.}

\item{value}{ANY. The R object to save in the cache.}

\item{cache_dir}{character. The cache directory. The default is
\code{cache_directory()}.}
}
\description{
Helper function for saving a file to a cache directory.
}

