% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.get.R, R/s3.put.R
\name{s3.get}
\alias{s3.get}
\alias{s3.put}
\title{Fetch an R object from an S3 path.}
\usage{
s3.get(path, bucket.location = "US", verbose = FALSE, debug = FALSE,
  cache = TRUE)

s3.put(x, path, name, bucket.location = "US", debug = FALSE,
  check_exists = TRUE, num_retries = getOption("s3mpi.num_retries", 0),
  backoff = 2^seq(2, num_retries + 1), max_backoff = 128)
}
\arguments{
\item{path}{character. A full S3 path.}

\item{bucket.location}{character. Usually \code{"US"}.}

\item{verbose}{logical. If \code{TRUE}, the \code{s3cmd}
utility verbose flag will be set.}

\item{debug}{logical. If \code{TRUE}, the \code{s3cmd}
utility debug flag will be set.}

\item{cache}{logical. If \code{TRUE}, an LRU in-memory cache will be referenced.}

\item{x}{ANY. R object to store to S3.}

\item{name}{character.}

\item{check_exists}{logical. Whether or not to check if an object already exists at the specificed location.}

\item{num_retries}{numeric. the number of times to retry uploading.}

\item{backoff}{numeric. Vector, with each element in seconds, describing the
exponential backoff to be used in conjunction with the num_retries argument.
Number of elements must equal num_retries. Defaults to 4, 8, 16, 32, etc.}

\item{max_backoff}{numeric. Number describing the maximum seconds s3mpi will sleep
prior to retrying an upload. Defaults to 128 seconds.}
}
\value{
For \code{s3.get}, the R object stored in RDS format on S3 in the \code{path}.
   For \code{s3.put}, the system exit code from running the \code{s3cmd}
   command line tool to perform the upload.
}
\description{
Fetch an R object from an S3 path.
}

