% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3exists.R
\name{s3exists}
\alias{s3exists}
\title{Determine whether object exists on S3.}
\usage{
s3exists(name, path = s3path())
}
\arguments{
\item{name}{string. Name of file to look for}

\item{path}{string. Path to file.  If missing, the entire s3 path must be provided in name.}
}
\description{
Test whether or not the given object exists at the
give S3 path.
}
\examples{
\dontrun{
s3exists("my/key") # Will look in bucket given by getOption("s3mpi.path")
  # For example, if this option is "s3://mybucket/", then this query
  # will check for existence of the \\code{s3://mybucket/my/key} S3 path.

s3exists("my/key", "s3://anotherbucket/") # We can of course change the bucket.
}
}

