% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3read.r
\name{s3read}
\alias{s3read}
\title{Read an R object in S3 by key}
\usage{
s3read(name, path = s3path(), cache = TRUE, serialize = TRUE, ...)
}
\arguments{
\item{name}{character. The key to grab from S3.}

\item{path}{character. The location of your S3 bucket as a prefix to \code{name},
for example, \code{"s3://mybucket/"} or \code{"s3://mybucket/myprefix/"}.}

\item{cache}{logical. If true, use the local s3cache if available.
If false, do not use cache. By default, \code{TRUE}. Note this will
consume local disk space for objects that have been \code{\link{s3read}}.}

\item{serialize}{logical. If true, use \code{s3normalize} to serialize the model object.}

\item{...}{Can be used internally to pass more arguments to \code{\link{s3.get}}.}
}
\description{
Any type of object that can be serialized as an RDS file
is capable of being read using this interface.
}
\details{
If you wish to read non-vanilla R objects, such as those
containing external pointers to C structures, see
\code{\link{s3normalize}}.
}
\examples{
\dontrun{
s3store(c(1,2,3), "test123")
print(s3read("test123"))
# [1] 1 2 3

s3store(function(x, y) { x + 2 * y }, "myfunc")
stopifnot(s3read("myfunc")(1, 2) == 5) # R can serialize closures!
}
}
\seealso{
\code{\link{s3store}}
}

